requirements.txt
easydict==1.13
einops==0.8.0
h5py==3.11.0
hydra-core==1.3.2
imageio==2.34.1
matplotlib==3.9.0
numpy==1.26.4
positional-encodings==6.0.3
torch==2.3.0
tqdm==4.66.4
vector-quantize-pytorch==1.14.24
wandb==0.17.0
metaworld @ git+https://github.com/Farama-Foundation/Metaworld.git@master#egg=metaworld
cmake==3.29.3